Added support for tables with .Disconnect method
Added support for tables with `.Disconnect` method, and is needed to work on signals exposed by a profile loaded from ProfileService. For example, a profile has a `MetaTagsUpdated` signal, which returns a table with a `.Disconnect` method rather than a `.Destroy` method and Maid won't clean it up upon calling `DoCleaning` or `Destroy` on the maid object. This commit fixes this issue. 

Also added `Maid.Cleanup` as a alias for `Maid.DoCleaning`, since `Maid.Cleanup` looks a lot clean as well and not a lot of people (including me) like to ruin my code's consistency.