GetService and GetController.
Removed GetService and GetController discrepancys.
Edited MoonWave documentation.
Added more info to errors

I tried to put in as much info as possible into the errors because being able to fix issues without having to google it is a good time saver. It also include links to the execution model and Knit Client info. It would be nice if all errors included the page of the funtion you're trying to run and also the OOS discord server invite link.